requirements.txt
arrow
cryptography
pyyaml
pytest
psycopg2
psycopg2-binary
pylint
pyjwt
python-jose-cryptodome
python-jose
Flask-Cors
flask
flask-wtf
flask-login
flask-sqlalchemy
flask_migrate
flask_script
gunicorn
requests
SQLALchemy
SQLAlchemy-utils
